+++ src/javascript/components/bootanimation.js
@@ -5,12 +5,13 @@ const BootAnimation = puffin.element(`
 
         &{
             padding:100px;
-            background:green;
+            background:white;
             position:fixed;
             top:0;
             left:0;
             min-height:100%;
             min-width:100%;
+            color:black;
         }
     
     `}">