+++ src/javascript/constructors/window.js
@@ -3,11 +3,13 @@ import WindowBody  from '../components/window/window'
 import WindowBackground  from '../components/window/background'
 
 function Window({
-    component
+    component,
+    height = "75%",
+    width = "80%"
 }){
     const randomID = Math.random()
     const WindowComponent = puffin.element(`
-        <div id="${randomID}" class="${puffin.style.css`
+        <div id="${randomID}"  class="${puffin.style.css`
             &{
                 min-height:100%;
                 min-width:100%;
@@ -19,7 +21,7 @@ function Window({
             }
         `}">
             <WindowBackground window="${randomID}"/>
-            <WindowBody>
+            <WindowBody style="height:${height};width:${width};">
                 <component/>
             </WindowBody>
         </div>