+++ src/javascript/defaults/menus.js
@@ -1,9 +1,11 @@
 import Menu from "../constructors/menu";
 import SettingsPage from '../defaults/settings'
 import Welcome from "./welcome";
+import Panel from '../constructors/panel'
+import Dialog from '../constructors/dialog'
 
 function loadDefaultMenus(){
-     new Menu({
+    new Menu({
          button:'File',
          list:[
             {
@@ -14,7 +16,7 @@ function loadDefaultMenus(){
             }
          ]
      })
-     new Menu({
+    new Menu({
         button:'Tools',
         list:[
             {
@@ -24,6 +26,27 @@ function loadDefaultMenus(){
             {
                 label:'Open Welcome',
                 action:Welcome.launch
+            }
+        ]
+    })
+    new Menu({
+        button:'Editor',
+        list:[
+            {
+                label:'New panel',
+                action:()=> new Panel()
+            }
+        ]
+    })
+    new Menu({
+        button:'Help',
+        list:[
+            {
+                label:'About',
+                action:()=> new Dialog({
+                    title:'About',
+                    content:'Graviton v2'
+                })
             }
         ]
     })